Refactor GridAsset, introduce openlayers - MEDT-3454
Now displaying images with openlayers.
