+++ mediathread/templates/base_new.html
@@ -56,6 +56,7 @@ A new base.html without the course context overwritten in the template.
 
         {% compress css %}
         <link rel="stylesheet" href='{% static "css/mediathread_new.css" %}' media="screen" />
+        <link rel="stylesheet" href='{% static "css/ol.css" %}' media="screen" />
         {% block css %}{% endblock %}
         {% endcompress %}
